refactor: cpe23 out of domain crate
