+++ cpe/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "cpe"
+version = "0.1.0"
+edition = "2021"
+
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
