+++ cpe/src/component.rs
